+++ compiler/src/iree/compiler/Dialect/LinalgExt/Utils/BUILD.bazel
@@ -16,14 +16,18 @@ iree_compiler_cc_library(
     name = "Utils",
     srcs = [
         "IndexingUtils.cpp",
+        "MatchUtils.cpp.cpp",
         "Utils.cpp",
     ],
     hdrs = [
         "IndexingUtils.h",
+        "MatchUtils.h"
         "Utils.h",
         "WinogradConstants.h",
     ],
     deps = [
+        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
+        "//compiler/src/iree/compiler/Codegen/Dialect/GPU/IR:IREEGPUDialect",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:AffineDialect",
         "@llvm-project//mlir:Analysis",
