+++ compiler/src/iree/compiler/Dialect/LinalgExt/Utils/CMakeLists.txt
@@ -15,10 +15,12 @@ iree_cc_library(
     Utils
   HDRS
     "IndexingUtils.h"
+    "MatchUtils.h"
     "Utils.h"
     "WinogradConstants.h"
   SRCS
     "IndexingUtils.cpp"
+    "MatchUtils.cpp"
     "Utils.cpp"
   DEPS
     LLVMSupport
@@ -26,6 +28,7 @@ iree_cc_library(
     MLIRAnalysis
     MLIRArithDialect
     MLIRComplexDialect
+    MLIRGPUDialect
     MLIRIR
     MLIRLinalgDialect
     MLIRLinalgInterfacesIncGenLib
@@ -35,6 +38,7 @@ iree_cc_library(
     MLIRTensorDialect
     MLIRTensorUtils
     MLIRTransformUtils
+    iree::compiler::Codegen::Dialect::Codegen::IR::IREECodegenDialect
   PUBLIC
 )
